scanner: Move ArtifactoryStorage to a sub-package / -directory
As more storage backends are about to come. Also move and rename the
HttpStorageTest to storages/ArtifactoryStorageTest as this test
effectively only tests the ArtifactoryStorage.
